video_kb_tool/requirements.txt
yt-dlp>=2025.1.0
faster-whisper>=1.1.0
openai>=1.30.0
rich>=13.7.0
python-slugify>=8.0.4
jieba>=0.42.1

streamlit>=1.44.0
